# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Хочу_То — Rental Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/listings:
    get:
      operationId: getUserListings
      tags: [users]
      summary: Get user's own listings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /listings/{id}/reviews:
    get:
      operationId: getListingReviews
      tags: [reviews]
      summary: Get reviews for a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [renter, owner, admin]
        phone:
          type: string
        avatar:
          type: string
        regionId:
          type: integer
        regionName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: string
        role:
          type: string
        regionName:
          type: string
        createdAt:
          type: string
        totalListings:
          type: integer
        totalBookings:
          type: integer
        rating:
          type: number
      required:
        - id
        - name
        - role
        - createdAt
        - totalListings
        - totalBookings
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
        regionId:
          type: integer
        role:
          type: string
          enum: [renter, owner]
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        pricePerDay:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        regionId:
          type: integer
        regionName:
          type: string
        photos:
          type: array
          items:
            type: string
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerAvatar:
          type: string
        ownerPhone:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        deposit:
          type: number
        createdAt:
          type: string
      required:
        - id
        - title
        - pricePerDay
        - categoryId
        - regionId
        - ownerId
        - isAvailable
        - createdAt
...
    Review:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
        createdAt:
          type: string
      required:
        - id
        - listingId
        - authorId
        - authorName
        - rating
        - createdAt
    CreateReviewRequest:
      type: object
      properties:
        listingId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
      required:
        - listingId
        - rating
...